ドロネー三角形分割関数の作成(未テスト)
git-svn-id: https://svn.osdn.net/svnroot/ongeo/trunk@30 c2f3b2dd-89f1-46ee-9904-f084b81f7106

+++ STLIO.cpp
@@ -37,7 +37,7 @@ void ONGEO_ReadLine(ON_BinaryArchive &ba, ON_String &str){
 		size_t cp = ba.CurrentPosition();
 		if (!ba.ReadChar(100, buf)){
 			bs = ba.CurrentPosition() - cp;
-			ba.SeekFromCurrentPosition(-bs);
+			ba.SeekFromCurrentPosition(-static_cast<int>(bs));
 			ba.ReadChar(bs, buf);
 		}
 		for (size_t i = 0; i < bs; ++i){
